feat: test L1 reorg

+++ taiko/helper.go
@@ -41,7 +41,7 @@ func WaitHeight(ctx context.Context, t *hivesim.T, client *ethclient.Client, f f
 
 func GetBlockHashByNumber(ctx context.Context, t *hivesim.T, cli *ethclient.Client, num *big.Int, needWait bool) common.Hash {
 	if needWait {
-		WaitHeight(ctx, t, cli, Greater(num.Uint64()-1))
+		WaitHeight(ctx, t, cli, Greater(int64(num.Uint64()-1)))
 	}
 	block, err := cli.BlockByNumber(ctx, num)
 	require.NoError(t, err)
@@ -143,13 +143,13 @@ func GenCommitDelayBlocks(t *hivesim.T, env *TestEnv) {
 	cnt := int(env.L1Constants.CommitDelayConfirmations.Uint64())
 	for i := 0; i < cnt; i++ {
 		env.L1Vault.CreateAccount(env.Context, l1.EthClient(t), big.NewInt(params.GWei))
-		WaitHeight(env.Context, t, l1.EthClient(t), Greater(curr+uint64(i)))
+		WaitHeight(env.Context, t, l1.EthClient(t), Greater(int64(curr)+int64(i)))
 	}
 }
 
-func Greater(want uint64) func(uint64) bool {
+func Greater(want int64) func(uint64) bool {
 	return func(get uint64) bool {
-		if get > want {
+		if int64(get) > want {
 			return true
 		}
 		return false
